% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zplot_function.R
\name{zplot_function}
\alias{zplot_function}
\title{A zplot_Function}
\usage{
zplot_function(x, mean, sd, tail)
}
\arguments{
\item{x, mean, sd, tail}{input the random variable x, its mean , sd and the tail probability of interest}
}
\description{
This function allows you find the z value and shows the p-value with a chart using the tigerstats package function (pnormGC).
}
\examples{
zplot_function(x,mean, sd, tail)
}
\keyword{zplot}
